* changed bucket addressing within a branch to use int instead of byte (underlying memory structures can still be bytes, but now they are unsigned) allowing for more than 127 buckets per branch (e.g. 169 in preflop)

+++ ao/bucket/abstraction/access/BucketSequencer.java
@@ -107,19 +107,20 @@ public class BucketSequencer
             Community  community)
     {
         CanonHole canonHole  = hole.asCanon();
-        byte      holeBucket = tree.getHole(canonHole.canonIndex());
+        int       holeBucket = tree.getHole(canonHole.canonIndex());
 
         Flop flop        = canonHole.addFlop(community);
-        byte flopBucket  = tree.getFlop( flop.canonIndex() );
+        int  flopBucket  = tree.getFlop( flop.canonIndex() );
 
         Turn turn        = flop.addTurn(community.turn());
-        byte turnBucket  = tree.getTurn( turn.canonIndex() );
+        int  turnBucket  = tree.getTurn( turn.canonIndex() );
 
         River river       = turn.addRiver(community.river());
-        byte  riverBucket = tree.getRiver( river.canonIndex() );
+        int  riverBucket = tree.getRiver( river.canonIndex() );
 
         return new byte[]{
-                holeBucket, flopBucket, turnBucket, riverBucket};
+                (byte) holeBucket, (byte) flopBucket,
+                (byte) turnBucket, (byte) riverBucket};
     }
